crypto/x509: Correct invalid verification of Name Constraints
As specified in RFC5280:

  The constraint MUST be specified as a fully qualified domain name and
  MAY specify a host or a domain.  Examples would be "host.example.com"
  and ".example.com".  When the constraint begins with a period, it MAY be
  expanded with one or more labels.  That is, the constraint
  ".example.com" is satisfied by both host.example.com and
  my.host.example.com.  However, the constraint ".example.com" is not
  satisfied by "example.com".  When the constraint does not begin with a
  period, it specifies a host.

The first conditional `opts.DNSName == domain` takes care of the case
where the constraint applies to a host; if the domain matches the
constraint exactly, then it is satisfied.

Now we are left with the case where the constraint is a domain (which
can match sub-domains). We can verify this with two checks:
- Is the first character of the constraint a period?
- Is the constraint a suffix of the domain in question?

The second check ensures that we are not matching part of a label in the
domain name.

This change simplifies the code, and makes it read just like the RFC.

The existing code exhibits an off-by-one error. It is effectively
checking to see if the character in the DNS name, right before the
matched suffix is a period. This is incorrect, because it doesn't
account for the period in the matched suffix, and instead looks at the
character *before* the period. (e.g. the 'b' in sub.example.com, if
domain = ".example.com"). We could simply remove the -1, but the code
would still be overly complex, considering we already verified the
period is present in opts.DNSName (by checking HasSuffix).